+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.20.0-rc
+
+- Added experimental `expand`, `filter`, `fields`, custom query and headers parameters support for the realtime subscriptions.
+    ```js
+    pb.collection("example").subscribe("*", (e) => {
+      ...
+    }, { filter: "someField > 10" });
+    ```
+    _This works only with PocketBase v0.20.0-rc._
+
+
 ## 0.19.0
 
 - Added `pb.filter(rawExpr, params?)` helper to construct a filter string with placeholder parameters populated from an object.